Move comment_case_runs view to testruns app (#913)
