+++ src/tests/core/test_views.py
@@ -7,11 +7,9 @@ from typing import List
 import pytest
 from django import test
 from django.contrib.auth.models import User
-from django.contrib.contenttypes.models import ContentType
 from django.core import serializers
 from django.core.exceptions import ValidationError
 from django.urls import reverse
-from django_comments.models import Comment
 from kobo.django.xmlrpc.models import XmlRpcLog
 from pytest_django import asserts
 
@@ -20,7 +18,6 @@ from tcms.core.templatetags.report_tags import percentage
 from tcms.management.models import TCMSEnvGroup
 from tcms.management.models import TCMSEnvProperty
 from tcms.testcases.forms import CaseAutomatedForm
-from tcms.testruns.models import TestCaseRun
 from tests import BaseCaseRun, BasePlanCase
 from tests import factories as f
 
@@ -97,67 +94,6 @@ class TestQuickSearch(BaseCaseRun):
     def test_404_when_missing_search_type(self):
         response = self.client.get(self.search_url, {"search_content": "python"})
         self.assert404(response)
-
-
-class TestCommentCaseRuns(BaseCaseRun):
-    """Test case for ajax.comment_case_runs"""
-
-    auto_login = True
-
-    @classmethod
-    def setUpTestData(cls):
-        super().setUpTestData()
-        cls.many_comments_url = reverse("caserun-comment-caseruns")
-
-    def test_refuse_if_missing_comment(self):
-        response = self.client.post(
-            self.many_comments_url, {"run": [self.case_run_1.pk, self.case_run_2.pk]}
-        )
-        self.assertJsonResponse(
-            response, {"message": "Comments needed"}, status_code=HTTPStatus.BAD_REQUEST
-        )
-
-    def test_refuse_if_missing_no_case_run_pk(self):
-        response = self.client.post(self.many_comments_url, {"comment": "new comment", "run": []})
-        self.assertJsonResponse(
-            response,
-            {"message": "No runs selected."},
-            status_code=HTTPStatus.BAD_REQUEST,
-        )
-
-        response = self.client.post(self.many_comments_url, {"comment": "new comment"})
-        self.assertJsonResponse(
-            response,
-            {"message": "No runs selected."},
-            status_code=HTTPStatus.BAD_REQUEST,
-        )
-
-    def test_refuse_if_passed_case_run_pks_not_exist(self):
-        response = self.client.post(
-            self.many_comments_url,
-            {"comment": "new comment", "run": [99999998, 1009900]},
-        )
-        self.assertJsonResponse(
-            response,
-            {"message": "No caserun found."},
-            status_code=HTTPStatus.BAD_REQUEST,
-        )
-
-    def test_add_comment_to_case_runs(self):
-        new_comment = "new comment"
-        response = self.client.post(
-            self.many_comments_url,
-            {"comment": new_comment, "run": [self.case_run_1.pk, self.case_run_2.pk]},
-        )
-        self.assertJsonResponse(response, {})
-
-        # Assert comments are added
-        case_run_ct = ContentType.objects.get_for_model(TestCaseRun)
-
-        for case_run_pk in (self.case_run_1.pk, self.case_run_2.pk):
-            comments = Comment.objects.filter(object_pk=case_run_pk, content_type=case_run_ct)
-            self.assertEqual(new_comment, comments[0].comment)
-            self.assertEqual(self.tester, comments[0].user)
 
 
 class TestGetForm(test.TestCase):
